Refactor API client infra, update Movies page UI
- Introduced WhatsOnApiClientBase for shared HTTP logic, logging, and query string handling
- Updated IMovieApiClient and IShowApiClient to remove CancellationToken and simplify signatures
- Refactored MovieApiClient and ShowApiClient to use new base class and require ILogger
- Changed PagedResult.TotalItemCount from long to int
- ResponseBase<T>.ErrorResult now returns strongly-typed ErrorResponse
- Replaced ApiErrorResponse with ErrorResponse in WhatsOn.Domain
- Moved WhatsOnApiOptions to Common.ApiClientBase
- Rewrote Movies.razor to use code-behind and GrabaUIPackage DataGrid for paging/sorting
- Improved Movies page state management and async loading
- Added Microsoft.AspNetCore.WebUtilities and GrabaUIPackage to .csproj
- Moved WhatsOn.Domain to its own solution folder
- Reformatted CSS files for consistent indentation

+++ WhatsOn.Domain/PagedResult.cs
@@ -6,6 +6,6 @@
 		public int PageNumber { get; init; }
 		public int PageSize { get; init; }
 		public int TotalPages { get; init; }
-		public long TotalItemCount { get; init; }
+		public int TotalItemCount { get; init; }
 	}
 }
